requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1
django-cors-headers==4.4.0
django-import-export==4.1.1
joblib==1.4.2
numpy==2.1.0
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
